mcp/gcalendar-mcp/pyproject.toml
[project]
name = "gcalendar-mcp"
version = "0.1.0"
description = "Multi-account Google Calendar MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.116.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
]

[project.scripts]
gcalendar-mcp = "gcalendar_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gcalendar_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
